Servant.API.ContentTypes

Description

A collection of basic Content-Types (also known as Internet Media Types, or MIME types). Additionally, this module provides classes that encapsulate how to serialize or deserialize values to or from a particular Content-Type.

Content-Types are used in ReqBody and the method combinators:

>>> type MyEndpoint = ReqBody '[JSON, PlainText] Book :> Put '[JSON, PlainText] Book

Meaning the endpoint accepts requests of Content-Type application/json or text/plain;charset-utf8, and returns data in either one of those formats (depending on the Accept header).

If you would like to support Content-Types beyond those provided here, then:

Declare a new data type with no constructors (e.g. data HTML).

Make an instance of it for Accept.

If you want to be able to serialize data *into* that Content-Type, make an instance of it for MimeRender.

If you want to be able to deserialize data *from* that Content-Type, make an instance of it for MimeUnrender.

Note that roles are reversed in servant-server and servant-client: to be able to serve (or even typecheck) a Get '[JSON, XML] MyData, you'll need to have the appropriate MimeRender instances in scope, whereas to query that endpoint with servant-client, you'll need a MimeUnrender instance in scope.

Provided Content-Types

data JSON

data PlainText

data FormUrlEncoded

data OctetStream

Building your own Content-Type

class Accept ctype where

Instances of Accept represent mimetypes. They are used for matching against the Accept HTTP header of the request, and for setting the Content-Type header of the response

Example:

>>> import Network.HTTP.Media ((//), (/:))
>>> data HTML
>>> :{
instance Accept HTML where
   contentType _ = "text" // "html" /: ("charset", "utf-8")
:}

Minimal complete definition

contentType | contentTypes

Methods

contentType :: Proxy ctype -> MediaType

contentTypes :: Proxy ctype -> NonEmpty MediaType

class Accept ctype => MimeRender ctype a where

Instantiate this class to register a way of serializing a type based on the Accept header.

Example:

data MyContentType

instance Accept MyContentType where
   contentType _ = "example" // "prs.me.mine" /: ("charset", "utf-8")

instance Show a => MimeRender MyContentType a where
   mimeRender _ val = pack ("This is MINE! " ++ show val)

type MyAPI = "path" :> Get '[MyContentType] Int

Methods

mimeRender :: Proxy ctype -> a -> ByteString

class Accept ctype => MimeUnrender ctype a where

Instantiate this class to register a way of deserializing a type based on the request's Content-Type header.

>>> import Network.HTTP.Media hiding (Accept)
>>> import qualified Data.ByteString.Lazy.Char8 as BSC
>>> data MyContentType = MyContentType String

>>> :{
instance Accept MyContentType where
   contentType _ = "example" // "prs.me.mine" /: ("charset", "utf-8")
:}

>>> :{
instance Read a => MimeUnrender MyContentType a where
   mimeUnrender _ bs = case BSC.take 12 bs of
     "MyContentType" -> return . read . BSC.unpack $ BSC.drop 12 bs
     _ -> Left "didn't start with the magic incantation"
:}

>>> type MyAPI = "path" :> ReqBody '[MyContentType] Int :> Get '[JSON] Int

Minimal complete definition

mimeUnrender | mimeUnrenderWithType

Methods

mimeUnrender :: Proxy ctype -> ByteString -> Either String a

mimeUnrenderWithType :: Proxy ctype -> MediaType -> ByteString -> Either String a

Variant which is given the actual MediaType provided by the other party.

In the most cases you don't want to branch based on the MediaType. See pr552 for a motivating example.

NoContent

data NoContent

A type for responses without content-body.

Constructors

NoContent

Internal

newtype AcceptHeader

Constructors

AcceptHeader ByteString

class AllMime list => AllCTRender (list :: [*]) a where

Methods

handleAcceptH :: Proxy list -> AcceptHeader -> a -> Maybe (ByteString, ByteString)

class AllCTUnrender (list :: [*]) a where

Minimal complete definition

canHandleCTypeH

Methods

canHandleCTypeH :: Proxy list -> ByteString -> Maybe (ByteString -> Either String a)

handleCTypeH :: Proxy list -> ByteString -> ByteString -> Maybe (Either String a)

class AllMime (list :: [*]) where

Methods

allMime :: Proxy list -> [MediaType]

class AllMime list => AllMimeRender (list :: [*]) a where

Methods

allMimeRender :: Proxy list -> a -> [(MediaType, ByteString)]

class AllMime list => AllMimeUnrender (list :: [*]) a where

Methods

allMimeUnrender :: Proxy list -> [(MediaType, ByteString -> Either String a)]

eitherDecodeLenient :: FromJSON a => ByteString -> Either String a

Deprecated: use eitherDecode instead

Deprecated: since aeson version 0.9 eitherDecode has lenient behavior.

canHandleAcceptH :: AllMime list => Proxy list -> AcceptHeader -> Bool