Servant.API.Description

Combinators

data Description (sym :: Symbol)

Add more verbose description for (part of) API.

Example:

>>> :{
type MyApi = Description
 "This comment is visible in multiple Servant interpretations \
 \and can be really long if necessary. \
 \Haskell multiline String support is not perfect \
 \but it's still very readable."
:> Get '[JSON] Book
:}

data Summary (sym :: Symbol)

Add a short summary for (part of) API.

Example:

>>> type MyApi = Summary "Get book by ISBN." :> "books" :> Capture "isbn" Text :> Get '[JSON] Book

Used as modifiers

type FoldDescription mods = FoldDescription' "" mods

Fold list of modifiers to extract description as a type-level String.

>>> :kind! FoldDescription '[]
FoldDescription '[] :: Symbol
= ""

>>> :kind! FoldDescription '[Required, Description "foobar", Lenient]
FoldDescription '[Required, Description "foobar", Lenient] :: Symbol
= "foobar"

type family FoldDescription' (acc :: Symbol) (mods :: [*]) :: Symbol where ...

Implementation of FoldDescription.

Equations

FoldDescription' acc '[] = acc

FoldDescription' acc (Description desc ': mods) = FoldDescription' desc mods

FoldDescription' acc (mod ': mods) = FoldDescription' acc mods

reflectDescription :: forall mods. KnownSymbol (FoldDescription mods) => Proxy mods -> String

Reflect description to the term level.

>>> reflectDescription (Proxy :: Proxy '[Required, Description "foobar", Lenient])
"foobar"