Servant.API.Generic

Description

Define servant servers from record types. Generics for the win.

The usage is simple, if you only need a collection of routes. First you define a record with field types prefixed by a parameter route:

data Routes route = Routes
    { _get :: route :- Capture "id" Int :> Get '[JSON] String
    , _put :: route :- ReqBody '[JSON] Int :> Put '[JSON] Bool
    }
  deriving (Generic)

You can get a Proxy of the server using

api :: Proxy (ToServantApi Routes)
api = genericApi (Proxy :: Proxy Routes)

Using genericApi is better as it checks that instances exists, i.e. you get better error messages than simply using Proxy value.

Note: in 0.14 series this module isn't re-exported from API.

Servant.API.Generic is based on servant-generic package by Patrick Chilton

Since: 0.14.1

Documentation

class GenericMode (mode :: k)

A class with a type family that applies an appropriate type family to the api parameter. For example, AsApi will leave api untouched, while AsServerT m will produce ServerT api m.

Associated Types

type (mode :: k) :- api infixl 0

type GenericServant (routes :: k -> Type) (mode :: k) = (GenericMode mode, Generic (routes mode), GServantProduct (Rep (routes mode)))

A constraint alias, for work with mode and routes.

type ToServant (routes :: k -> Type) (mode :: k) = GToServant (Rep (routes mode))

Turns a generic product type into a tree of :<|> combinators.

toServant :: forall {k} routes (mode :: k). GenericServant routes mode => routes mode -> ToServant routes mode

See ToServant, but at value-level.

fromServant :: forall {k} routes (mode :: k). GenericServant routes mode => ToServant routes mode -> routes mode

Inverse of toServant.

This can be used to turn generated values such as client functions into records.

You may need to provide a type signature for the output type (your record type).

AsApi

data AsApi

A type that specifies that an API record contains an API definition. Only useful at type-level.

type ToServantApi (routes :: Type -> Type) = ToServant routes AsApi

genericApi :: forall (routes :: Type -> Type). GenericServant routes AsApi => Proxy routes -> Proxy (ToServantApi routes)

Get a Proxy of an API type.

Utility

class GServantProduct (f :: k -> Type)

Minimal complete definition

gtoServant, gfromServant

re-exports

class Generic a

Representable types of kind *. This class is derivable in GHC with the DeriveGeneric flag on.

A Generic instance must satisfy the following laws:

from . to ≡ id
to . from ≡ id

Minimal complete definition

from, to

Associated Types

type Rep a :: Type -> Type

Generic representation type