Servant.API.HttpVersion

Documentation

| You can directly use the HttpVersion type from Network.HTTP.Types if your request handlers need it to compute a response. This would make the request handlers take an argument of type HttpVersion.

Example:

>>> type API = HttpVersion :> Get '[JSON] String

data HttpVersion

HTTP Version.

Note that the Show instance is intended merely for debugging.

Constructors

HttpVersion

Fields

httpMajor :: !Int

httpMinor :: !Int