Servant.API.Put

Documentation

data Put a

Endpoint for PUT requests, usually used to update a ressource. The type a is the type of the response body that's returned.

Example:

-- PUT /books/:isbn
-- with a Book as request body, returning the updated Book
type MyApi = "books" :> Capture "isbn" Text :> ReqBody Book :> Put Book