Servant.API.Stream

Documentation

data Stream (method :: k1) (status :: Nat) (framing :: *) (contentType :: *) (a :: *)

A Stream endpoint for a given method emits a stream of encoded values at a given Content-Type, delimited by a framing strategy. Type synonyms are provided for standard methods.

type StreamGet = Stream 'GET 200

type StreamPost = Stream 'POST 200

type StreamBody = StreamBody' '[]

A stream request body.

data StreamBody' (mods :: [*]) (framing :: *) (contentType :: *) (a :: *)

Source

SourceIO are equivalent to some *source* in streaming libraries.

type SourceIO = SourceT IO

Stream endpoints may be implemented as producing a SourceIO chunk.

Clients reading from streaming endpoints can be implemented as consuming a SourceIO chunk.

class ToSourceIO chunk a | a -> chunk where

ToSourceIO is intended to be implemented for types such as Conduit, Pipe, etc. By implementing this class, all such streaming abstractions can be used directly as endpoints.

Methods

toSourceIO :: a -> SourceIO chunk

class FromSourceIO chunk a | a -> chunk where

FromSourceIO is intended to be implemented for types such as Conduit, Pipe, etc. By implementing this class, all such streaming abstractions can be used directly on the client side for talking to streaming endpoints.

Methods

fromSourceIO :: SourceIO chunk -> IO a

Auxiliary classes

class SourceToSourceIO m where

Auxiliary class for ToSourceIO x (SourceT m x) instance.

Methods

sourceToSourceIO :: SourceT m a -> SourceT IO a

Framing

class FramingRender strategy where

The FramingRender class provides the logic for emitting a framing strategy. The strategy transforms a SourceT m a into SourceT m ByteString, therefore it can prepend, append and intercalate framing structure around chunks.

Note: as the Monad m is generic, this is pure transformation.

Methods

framingRender :: Monad m => Proxy strategy -> (a -> ByteString) -> SourceT m a -> SourceT m ByteString

class FramingUnrender strategy where

The FramingUnrender class provides the logic for parsing a framing strategy.

Methods

framingUnrender :: Monad m => Proxy strategy -> (ByteString -> Either String a) -> SourceT m ByteString -> SourceT m a

Strategies

data NoFraming

A framing strategy that does not do any framing at all, it just passes the input data This will be used most of the time with binary data, such as files

data NewlineFraming

A simple framing strategy that has no header, and inserts a newline character after each frame. This assumes that it is used with a Content-Type that encodes without newlines (e.g. JSON).

data NetstringFraming

The netstring framing strategy as defined by djb: http://cr.yp.to/proto/netstrings.txt

Any string of 8-bit bytes may be encoded as [len]":"[string]",". Here [string] is the string and [len] is a nonempty sequence of ASCII digits giving the length of [string] in decimal. The ASCII digits are 30 for 0, 31 for 1, and so on up through 39 for 9. Extra zeros at the front of [len] are prohibited: [len] begins with 30 exactly when [string] is empty.

For example, the string "hello world!" is encoded as 32 3a 68 65 6c 6c 6f 20 77 6f 72 6c 64 21 2c, i.e., "12:hello world!,". The empty string is encoded as "0:,".