Servant.API.Vault

Documentation

| Use Vault in your API types to provide access to the Vault of the request, which is a location shared by middlewares and applications to store arbitrary data. See vault for more details on how to actually use the vault in your handlers

Example:

>>> type API = Vault :> Get '[JSON] String

type Vault = Vault RealWorld

A persistent store for values of arbitrary types.

This variant is the simplest and creates keys in the IO monad. See the module Data.Vault.ST if you want to use it with the ST monad instead.