Servant.Common.Text

Documentation

class FromText a where

For getting values from url captures and query string parameters

Methods

fromText :: Text -> Maybe a

class ToText a where

For putting values in paths and query string parameters

Methods

toText :: a -> Text