Servant.QQ

Description

QuasiQuoting utilities for API types.

sitemap allows you to write your type in a very natural way:

[sitemap|
PUT        hello                 String -> ()
POST       hello/p:Int           String -> ()
GET        hello/?name:String    Int
|]

Will generate:

       "hello" :> ReqBody String :> Put ()
  :<|> "hello" :> Capture "p" Int :> ReqBody String :> Post ()
  :<|> "hello" :> QueryParam "name" String :> Get Int

Note the / before a QueryParam!

Documentation

class ExpSYM repr' repr | repr -> repr', repr' -> repr where

Finally-tagless encoding for our DSL. Keeping repr' and repr distinct when writing functions with an ExpSYM context ensures certain invariants (for instance, that there is only one of get, post, put, and delete in a value), but sometimes requires a little more work.

Methods

lit :: String -> repr' -> repr

capture :: String -> String -> repr -> repr

reqBody :: String -> repr -> repr

queryParam :: String -> String -> repr -> repr

conj :: repr' -> repr -> repr

get :: String -> repr

post :: String -> repr

put :: String -> repr

delete :: String -> repr

(>:) :: Type -> Type -> Type infixr 6

parseMethod :: ExpSYM repr' repr => Parser (String -> repr)

parseUrlSegment :: ExpSYM repr repr => Parser (repr -> repr)

parseUrl :: ExpSYM repr repr => Parser (repr -> repr)

data Typ

Constructors

Val String

ReqArgVal String String

parseTyp :: Parser Typ

parseEntry :: ExpSYM repr repr => Parser repr

blockComment :: Parser ()

inlineComment :: Parser ()

eol :: Parser String

eols :: Parser ()

parseAll :: Parser Type

sitemap :: QuasiQuoter

The sitemap QuasiQuoter.

...var:type... becomes a capture

.../?var:type becomes a query parameter

method ... typ becomes a method returning typ

method ... typ1 -> typ2 becomes a method with request body of typ1 and returning typ2

Comments are allowed, and have the standard Haskell format

-- for inline

{- ... -} for block