Servant.Utils.Enter

Description

Deprecated: Use hoistServer or hoistServerWithContext from servant-server

Documentation

type family Entered m n api where ...

Helper type family to state the Enter symmetry.

Equations

Entered m n (a -> api) = a -> Entered m n api

Entered m n (m a) = n a

Entered m n (api1 :<|> api2) = Entered m n api1 :<|> Entered m n api2

Entered m n (Tagged m a) = Tagged n a

class (Entered m n typ ~ ret, Entered n m ret ~ typ) => Enter typ m n ret | typ m n -> ret, ret m n -> typ, ret typ m -> n, ret typ n -> m where

Minimal complete definition

enter

Methods

enter :: (m :~> n) -> typ -> ret

Map the leafs of an API type.

Servant combinators

Leaf instances

liftNat :: (MonadTrans t, Monad m) => m :~> t m

Like lift.

runReaderTNat :: r -> ReaderT r m :~> m

evalStateTLNat :: Monad m => s -> StateT s m :~> m

evalStateTSNat :: Monad m => s -> StateT s m :~> m

logWriterTSNat :: MonadIO m => (w -> IO ()) -> WriterT w m :~> m

Log the contents of WriterT with the function provided as the first argument, and return the value of the WriterT computation

logWriterTLNat :: MonadIO m => (w -> IO ()) -> WriterT w m :~> m

Like logWriterTSNat, but for lazy WriterT.

hoistNat :: (MFunctor t, Monad m) => (m :~> n) -> t m :~> t n

Like mmorph's hoist.

embedNat :: (MMonad t, Monad n) => (m :~> t n) -> t m :~> t n

Like mmorph's embed.

squashNat :: (Monad m, MMonad t) => t (t m) :~> t m

Like mmorph's squash.

generalizeNat :: Applicative m => Identity :~> m

Like mmorph's generalize.

natural-transformation re-exports

newtype (k :~> f) g :: forall k. (k -> *) -> (k -> *) -> * infixr 0

A natural transformation suitable for storing in a container.

Constructors

NT

Fields

($$) :: (~>) k f g