Servant.API

Combinators

module Servant.API.Sub

Type-level combinator for expressing subrouting: :>

module Servant.API.Alternative

Type-level combinator for alternative endpoints: :<|>

module Servant.API.Empty

Type-level combinator for an empty API: EmptyAPI

module Servant.API.Modifiers

Type-level modifiers for QueryParam, Header and ReqBody.

Accessing information from the request

module Servant.API.Capture

Capturing parts of the url path as parsed values: Capture and CaptureAll

module Servant.API.Header

Retrieving specific headers from the request

module Servant.API.HttpVersion

Retrieving the HTTP version of the request

module Servant.API.QueryParam

Retrieving parameters from the query string of the URI: QueryParam

module Servant.API.ReqBody

Accessing the request body as a JSON-encoded type: ReqBody

module Servant.API.RemoteHost

Retrieving the IP of the client

module Servant.API.IsSecure

Is the request made through HTTPS?

module Servant.API.Vault

Access the location for arbitrary data to be shared by applications and middleware

module Servant.API.WithNamedContext

Access context entries in combinators in servant-server

Actual endpoints, distinguished by HTTP method

module Servant.API.Verbs

Streaming endpoints, distinguished by HTTP method

module Servant.API.Stream

Authentication

module Servant.API.BasicAuth

Endpoints description

module Servant.API.Description

Content Types

module Servant.API.ContentTypes

Serializing and deserializing types based on Accept and Content-Type headers.

Response Headers

module Servant.API.ResponseHeaders

Untyped endpoints

module Servant.API.Raw

Plugging in a wai Application, serving directories

FromHttpApiData and ToHttpApiData

module Web.HttpApiData

Classes and instances for types that can be converted to and from HTTP API data.

Experimental modules

module Servant.API.Experimental.Auth

General Authentication

Utilities

module Servant.Utils.Links

Type-safe internal URIs

Re-exports

type family If k (cond :: Bool) (tru :: k) (fls :: k) :: k where ...

Type-level If. If True a b ==> a; If False a b ==> b

Equations

If k True tru fls = tru

If k False tru fls = fls

data SBool b :: Bool -> * where

Constructors

STrue :: SBool True

SFalse :: SBool False

class SBoolI b where

Minimal complete definition

sbool

Methods

sbool :: SBool b