Servant.API

Combinators

module Servant.API.Sub

Type-level combinator for expressing subrouting: :>

module Servant.API.Alternative

Type-level combinator for alternative endpoints: :<|>

Accessing information from the request

module Servant.API.Capture

Capturing parts of the url path as parsed values: Capture

module Servant.API.Header

Retrieving specific headers from the request

module Servant.API.HttpVersion

Retrieving the HTTP version of the request

module Servant.API.QueryParam

Retrieving parameters from the query string of the URI: QueryParam

module Servant.API.ReqBody

Accessing the request body as a JSON-encoded type: ReqBody

module Servant.API.RemoteHost

Retrieving the IP of the client

module Servant.API.IsSecure

Is the request made through HTTPS?

module Servant.API.Vault

Access the location for arbitrary data to be shared by applications and middleware

Actual endpoints, distinguished by HTTP method

module Servant.API.Get

GET requests

module Servant.API.Post

POST requests

module Servant.API.Delete

DELETE requests

module Servant.API.Put

PUT requests

module Servant.API.Patch

PATCH requests

Content Types

module Servant.API.ContentTypes

Serializing and deserializing types based on Accept and Content-Type headers.

Response Headers

module Servant.API.ResponseHeaders

Untyped endpoints

module Servant.API.Raw

Plugging in a wai Application, serving directories

FromHttpApiData and ToHttpApiData

Classes and instances for types that can be converted to and from HTTP API data.

Utilities

module Servant.Utils.Links

Type-safe internal URIs