Servant.API

Combinators

data (path :: k) :> (a :: *) infixr 4

The contained API (second argument) can be found under ("/" ++ path) (path being the first argument).

Example:

>>> -- GET /hello/world
>>> -- returning a JSON encoded World value
>>> type MyApi = "hello" :> "world" :> Get '[JSON] World

Type-level combinator for expressing subrouting: :>

data a :<|> b infixr 3

Union of two APIs, first takes precedence in case of overlap.

Example:

>>> :{
type MyApi = "books" :> Get '[JSON] [Book] -- GET /books
       :<|> "books" :> ReqBody '[JSON] Book :> Post '[JSON] () -- POST /books
:}

Constructors

a :<|> b infixr 3

Type-level combinator for alternative endpoints: :<|>

data EmptyAPI

An empty API: one which serves nothing. Morally speaking, this should be the unit of :<|>. Implementors of interpretations of API types should treat EmptyAPI as close to the unit as possible.

Constructors

EmptyAPI

Type-level combinator for an empty API: EmptyAPI

data Strict

Strictly parsed argument. Not wrapped.

data Lenient

Leniently parsed argument, i.e. parsing never fail. Wrapped in Either Text.

data Optional

Optional argument. Wrapped in Maybe.

data Required

Required argument. Not wrapped.

Type-level modifiers for QueryParam, ResponseHeader and ReqBody.

Accessing information from the request

type Capture = Capture' '[]

Capture a value from the request path under a certain type a.

Example:

>>> -- GET /books/:isbn
>>> type MyApi = "books" :> Capture "isbn" Text :> Get '[JSON] Book

data Capture' (mods :: [*]) (sym :: Symbol) (a :: *)

Capture which can be modified. For example with Description.

data CaptureAll (sym :: Symbol) (a :: *)

Capture all remaining values from the request path under a certain type a.

Example:

>>> -- GET /src/*
>>> type MyAPI = "src" :> CaptureAll "segments" Text :> Get '[JSON] SourceFile

Capturing parts of the url path as parsed values: Capture and CaptureAll

type Header = Header' '[Optional, Strict]

Extract the given header's value as a value of type a. I.e. header sent by client, parsed by server.

Example:

>>> newtype Referer = Referer Text deriving (Eq, Show)
>>> 
>>> -- GET /view-my-referer
>>> type MyApi = "view-my-referer" :> Header "from" Referer :> Get '[JSON] Referer

data Header' (mods :: [*]) (sym :: Symbol) (a :: *)

Retrieving specific headers from the request

data HttpVersion

HTTP Version.

Note that the Show instance is intended merely for debugging.

Constructors

HttpVersion

Fields

httpMajor :: !Int

httpMinor :: !Int

Retrieving the HTTP version of the request

type QueryParam = QueryParam' '[Optional, Strict]

Lookup the value associated to the sym query string parameter and try to extract it as a value of type a.

Example:

>>> -- /books?author=<author name>
>>> type MyApi = "books" :> QueryParam "author" Text :> Get '[JSON] [Book]

data QueryFlag (sym :: Symbol)

Lookup a potentially value-less query string parameter with boolean semantics. If the param sym is there without any value, or if it's there with value "true" or "1", it's interpreted as True. Otherwise, it's interpreted as False.

Example:

>>> -- /books?published
>>> type MyApi = "books" :> QueryFlag "published" :> Get '[JSON] [Book]

data QueryParam' (mods :: [*]) (sym :: Symbol) (a :: *)

QueryParam which can be Required, Lenient, or modified otherwise.

data QueryParams (sym :: Symbol) (a :: *)

Lookup the values associated to the sym query string parameter and try to extract it as a value of type [a]. This is typically meant to support query string parameters of the form param[]=val1&param[]=val2 and so on. Note that servant doesn't actually require the []s and will fetch the values just fine with param=val1&param=val2, too.

Example:

>>> -- /books?authors[]=<author1>&authors[]=<author2>&...
>>> type MyApi = "books" :> QueryParams "authors" Text :> Get '[JSON] [Book]

Retrieving parameters from the query string of the URI: QueryParam

data Fragment (a :: *)

Document the URI fragment in API. Useful in combination with Link.

Example:

>>> -- /post#TRACKING
>>> type MyApi = "post" :> Fragment Text :> Get '[JSON] Tracking

Documenting the fragment of the URI: Fragment

type ReqBody = ReqBody' '[Required, Strict]

Extract the request body as a value of type a.

Example:

>>> -- POST /books
>>> type MyApi = "books" :> ReqBody '[JSON] Book :> Post '[JSON] Book

data ReqBody' (mods :: [*]) (contentTypes :: [*]) (a :: *)

Note: ReqBody' is always Required.

Accessing the request body as a JSON-encoded type: ReqBody

data RemoteHost

Provides access to the host or IP address from which the HTTP request was sent.

Retrieving the IP of the client

data IsSecure

Was this request made over an SSL connection?

Note that this value will not tell you if the client originally made this request over SSL, but rather whether the current connection is SSL. The distinction lies with reverse proxies. In many cases, the client will connect to a load balancer over SSL, but connect to the WAI handler without SSL. In such a case, the handlers would get NotSecure, but from a user perspective, there is a secure connection.

Constructors

Secure
the connection to the server is secure (HTTPS)

NotSecure
the connection to the server is not secure (HTTP)

Is the request made through HTTPS?

type Vault = Vault RealWorld

A persistent store for values of arbitrary types.

This variant is the simplest and creates keys in the IO monad. See the module Data.Vault.ST if you want to use it with the ST monad instead.

Access the location for arbitrary data to be shared by applications and middleware

data WithNamedContext (name :: Symbol) (subContext :: [*]) subApi

WithNamedContext names a specific tagged context to use for the combinators in the API. (See also in servant-server, Servant.Server.Context.) For example:

type UseNamedContextAPI = WithNamedContext "myContext" '[String] (
    ReqBody '[JSON] Int :> Get '[JSON] Int)

Both the ReqBody and Get combinators will use the WithNamedContext with type tag "myContext" as their context.

Contexts are only relevant for servant-server.

For more information, see the tutorial.

Access context entries in combinators in servant-server

data WithResource res

Access a managed resource scoped to a single request

Actual endpoints, distinguished by HTTP method

type Get = Verb 'GET 200

GET with 200 status code.

type Put = Verb 'PUT 200

PUT with 200 status code.

data StdMethod

HTTP standard method (as defined by RFC 2616, and PATCH which is defined by RFC 5789).

Since: http-types-0.2.0

Constructors

GET

POST

HEAD

PUT

DELETE

TRACE

CONNECT

OPTIONS

PATCH
Since: http-types-0.8.0

type Delete = Verb 'DELETE 200

DELETE with 200 status code.

type DeleteAccepted = Verb 'DELETE 202

DELETE with 202 status code.

type DeleteNoContent = NoContentVerb 'DELETE

DELETE with 204 status code.

type DeleteNonAuthoritative = Verb 'DELETE 203

DELETE with 203 status code.

type GetAccepted = Verb 'GET 202

GET with 202 status code.

type GetNoContent = NoContentVerb 'GET

GET with 204 status code.

type GetNonAuthoritative = Verb 'GET 203

GET with 203 status code.

type GetPartialContent = Verb 'GET 206

GET with 206 status code.

type GetResetContent = Verb 'GET 205

GET with 205 status code.

data NoContentVerb (method :: k1)

NoContentVerb is a specific type to represent NoContent responses. It does not require either a list of content types (because there's no content) or a status code (because it should always be 204).

type Patch = Verb 'PATCH 200

PATCH with 200 status code.

type PatchAccepted = Verb 'PATCH 202

PATCH with 202 status code.

type PatchNoContent = NoContentVerb 'PATCH

PATCH with 204 status code.

type PatchNonAuthoritative = Verb 'PATCH 203

PATCH with 203 status code.

type Post = Verb 'POST 200

POST with 200 status code.

type PostAccepted = Verb 'POST 202

POST with 202 status code.

type PostCreated = Verb 'POST 201

POST with 201 status code.

type PostNoContent = NoContentVerb 'POST

POST with 204 status code.

type PostNonAuthoritative = Verb 'POST 203

POST with 203 status code.

type PostResetContent = Verb 'POST 205

POST with 205 status code.

type PutAccepted = Verb 'PUT 202

PUT with 202 status code.

type PutCreated = Verb 'PUT 201

PUT with 201 status code.

type PutNoContent = NoContentVerb 'PUT

PUT with 204 status code.

type PutNonAuthoritative = Verb 'PUT 203

PUT with 203 status code.

class ReflectMethod a where

Methods

reflectMethod :: Proxy a -> Method

data Verb (method :: k1) (statusCode :: Nat) (contentTypes :: [*]) (a :: *)

Verb is a general type for representing HTTP verbs (a.k.a. methods). For convenience, type synonyms for each verb with a 200 response code are provided, but you are free to define your own:

>>> type Post204 contentTypes a = Verb 'POST 204 contentTypes a

type family Unique xs :: Constraint where ...

Check whether all values in a type-level list are distinct. This will throw a nice error if there are any duplicate elements in the list.

Equations

Unique xs = If (Nubbed xs == 'True) (() :: Constraint) (TypeError (DuplicateElementError xs))

type Union = NS I

data UVerb (method :: StdMethod) (contentTypes :: [*]) (as :: [*])

A variant of Verb that can have any of a number of response values and status codes.

FUTUREWORK: it would be nice to make Verb a special case of UVerb, and only write instances for HasServer etc. for the latter, getting them for the former for free. Something like:

type Verb method statusCode contentTypes a = UVerb method contentTypes [WithStatus statusCode a]

Backwards compatibility is tricky, though: this type alias would mean people would have to use respond instead of pure or return, so all old handlers would have to be rewritten.

class KnownStatus (StatusOf a) => HasStatus (a :: *)

Associated Types

type StatusOf (a :: *) :: Nat

type IsMember (a :: u) (as :: [u]) = (Unique as, CheckElemIsMember a as, UElem a as)

type family Statuses (as :: [*]) :: [Nat]

type family Statuses (as :: [*]) :: [Nat]

newtype WithStatus (k :: Nat) a

A simple newtype wrapper that pairs a type with its status code. It implements all the content types that Servant ships with by default.

Constructors

WithStatus a

inject :: UElem x xs => f x -> NS f xs

statusOf :: forall a proxy. HasStatus a => proxy a -> Status

Sub-APIs defined as records of routes

data NamedRoutes (api :: * -> *)

Combinator for embedding a record of named routes into a Servant API type.

class GenericMode mode

A class with a type family that applies an appropriate type family to the api parameter. For example, AsApi will leave api untouched, while AsServerT m will produce ServerT api m.

Associated Types

type mode :- api :: * infixl 0

data AsApi

A type that specifies that an API record contains an API definition. Only useful at type-level.

class GServantProduct f

Minimal complete definition

gtoServant, gfromServant

type GenericServant routes mode = (GenericMode mode, Generic (routes mode), GServantProduct (Rep (routes mode)))

A constraint alias, for work with mode and routes.

type ToServant routes mode = GToServant (Rep (routes mode))

Turns a generic product type into a tree of :<|> combinators.

type ToServantApi routes = ToServant routes AsApi

fromServant :: GenericServant routes mode => ToServant routes mode -> routes mode

Inverse of toServant.

This can be used to turn generated values such as client functions into records.

You may need to provide a type signature for the output type (your record type).

genericApi :: GenericServant routes AsApi => Proxy routes -> Proxy (ToServantApi routes)

Get a Proxy of an API type.

toServant :: GenericServant routes mode => routes mode -> ToServant routes mode

See ToServant, but at value-level.

Streaming endpoints, distinguished by HTTP method

data Stream (method :: k1) (status :: Nat) (framing :: *) (contentType :: *) (a :: *)

A Stream endpoint for a given method emits a stream of encoded values at a given Content-Type, delimited by a framing strategy. Type synonyms are provided for standard methods.

class FramingRender strategy where

The FramingRender class provides the logic for emitting a framing strategy. The strategy transforms a SourceT m a into SourceT m ByteString, therefore it can prepend, append and intercalate framing structure around chunks.

Note: as the Monad m is generic, this is pure transformation.

Methods

framingRender :: Monad m => Proxy strategy -> (a -> ByteString) -> SourceT m a -> SourceT m ByteString

class FramingUnrender strategy where

The FramingUnrender class provides the logic for parsing a framing strategy.

Methods

framingUnrender :: Monad m => Proxy strategy -> (ByteString -> Either String a) -> SourceT m ByteString -> SourceT m a

class FromSourceIO chunk a | a -> chunk where

FromSourceIO is intended to be implemented for types such as Conduit, Pipe, etc. By implementing this class, all such streaming abstractions can be used directly on the client side for talking to streaming endpoints.

Methods

fromSourceIO :: SourceIO chunk -> IO a

data NetstringFraming

The netstring framing strategy as defined by djb: http://cr.yp.to/proto/netstrings.txt

Any string of 8-bit bytes may be encoded as [len]":"[string]",". Here [string] is the string and [len] is a nonempty sequence of ASCII digits giving the length of [string] in decimal. The ASCII digits are 30 for 0, 31 for 1, and so on up through 39 for 9. Extra zeros at the front of [len] are prohibited: [len] begins with 30 exactly when [string] is empty.

For example, the string "hello world!" is encoded as 32 3a 68 65 6c 6c 6f 20 77 6f 72 6c 64 21 2c, i.e., "12:hello world!,". The empty string is encoded as "0:,".

data NewlineFraming

A simple framing strategy that has no header, and inserts a newline character after each frame. This assumes that it is used with a Content-Type that encodes without newlines (e.g. JSON).

data NoFraming

A framing strategy that does not do any framing at all, it just passes the input data This will be used most of the time with binary data, such as files

type SourceIO = SourceT IO

Stream endpoints may be implemented as producing a SourceIO chunk.

Clients reading from streaming endpoints can be implemented as consuming a SourceIO chunk.

type StreamBody = StreamBody' '[]

A stream request body.

data StreamBody' (mods :: [*]) (framing :: *) (contentType :: *) (a :: *)

type StreamGet = Stream 'GET 200

type StreamPost = Stream 'POST 200

class ToSourceIO chunk a | a -> chunk where

ToSourceIO is intended to be implemented for types such as Conduit, Pipe, etc. By implementing this class, all such streaming abstractions can be used directly as endpoints.

Methods

toSourceIO :: a -> SourceIO chunk

Authentication

data BasicAuth (realm :: Symbol) (userData :: *)

Combinator for Basic Access Authentication.

IMPORTANT*: Only use Basic Auth over HTTPS! Credentials are not hashed or encrypted. Note also that because the same credentials are sent on every request, Basic Auth is not as secure as some alternatives. Further, the implementation in servant-server does not protect against some types of timing attacks.

In Basic Auth, username and password are base64-encoded and transmitted via the Authorization header. Handshakes are not required, making it relatively efficient.

data BasicAuthData

A simple datatype to hold data required to decorate a request

Constructors

BasicAuthData

Fields

basicAuthUsername :: !ByteString

basicAuthPassword :: !ByteString

Endpoints description

data Description (sym :: Symbol)

Add more verbose description for (part of) API.

Example:

>>> :{
type MyApi = Description
 "This comment is visible in multiple Servant interpretations \
 \and can be really long if necessary. \
 \Haskell multiline String support is not perfect \
 \but it's still very readable."
:> Get '[JSON] Book
:}

data Summary (sym :: Symbol)

Add a short summary for (part of) API.

Example:

>>> type MyApi = Summary "Get book by ISBN." :> "books" :> Capture "isbn" Text :> Get '[JSON] Book

Content Types

class Accept ctype where

Instances of Accept represent mimetypes. They are used for matching against the Accept HTTP header of the request, and for setting the Content-Type header of the response

Example:

>>> import Network.HTTP.Media ((//), (/:))
>>> data HTML
>>> :{
instance Accept HTML where
   contentType _ = "text" // "html" /: ("charset", "utf-8")
:}

Minimal complete definition

contentType | contentTypes

Methods

contentType :: Proxy ctype -> MediaType

contentTypes :: Proxy ctype -> NonEmpty MediaType

data FormUrlEncoded

data JSON

class Accept ctype => MimeRender ctype a where

Instantiate this class to register a way of serializing a type based on the Accept header.

Example:

data MyContentType

instance Accept MyContentType where
   contentType _ = "example" // "prs.me.mine" /: ("charset", "utf-8")

instance Show a => MimeRender MyContentType a where
   mimeRender _ val = pack ("This is MINE! " ++ show val)

type MyAPI = "path" :> Get '[MyContentType] Int

Methods

mimeRender :: Proxy ctype -> a -> ByteString

class Accept ctype => MimeUnrender ctype a where

Instantiate this class to register a way of deserializing a type based on the request's Content-Type header.

>>> import Network.HTTP.Media hiding (Accept)
>>> import qualified Data.ByteString.Lazy.Char8 as BSC
>>> data MyContentType = MyContentType String

>>> :{
instance Accept MyContentType where
   contentType _ = "example" // "prs.me.mine" /: ("charset", "utf-8")
:}

>>> :{
instance Read a => MimeUnrender MyContentType a where
   mimeUnrender _ bs = case BSC.take 12 bs of
     "MyContentType" -> return . read . BSC.unpack $ BSC.drop 12 bs
     _ -> Left "didn't start with the magic incantation"
:}

>>> type MyAPI = "path" :> ReqBody '[MyContentType] Int :> Get '[JSON] Int

Minimal complete definition

mimeUnrender | mimeUnrenderWithType

Methods

mimeUnrender :: Proxy ctype -> ByteString -> Either String a

mimeUnrenderWithType :: Proxy ctype -> MediaType -> ByteString -> Either String a

Variant which is given the actual MediaType provided by the other party.

In the most cases you don't want to branch based on the MediaType. See pr552 for a motivating example.

data NoContent

A type for responses without content-body.

Constructors

NoContent

data OctetStream

data PlainText

Serializing and deserializing types based on Accept and Content-Type headers.

Response Headers

class AddHeader h v orig new | h v orig -> new, new -> h, new -> v, new -> orig

Minimal complete definition

addOptionalHeader

class BuildHeadersTo hs where

Methods

buildHeadersTo :: [Header] -> HList hs

class GetHeaders ls where

Methods

getHeaders :: ls -> [Header]

data HList a where

Constructors

HNil :: HList '[]

HCons :: ResponseHeader h x -> HList xs -> HList (Header h x ': xs)

class HasResponseHeader h a headers

Minimal complete definition

hlistLookupHeader

data Headers ls a

Response Header objects. You should never need to construct one directly. Instead, use addOptionalHeader.

Constructors

Headers

Fields

getResponse :: a

The underlying value of a Headers

getHeadersHList :: HList ls

HList of headers.

data ResponseHeader (sym :: Symbol) a

Constructors

Header a

MissingHeader

UndecodableHeader ByteString

addHeader :: AddHeader h v orig new => v -> orig -> new

addHeader adds a header to a response. Note that it changes the type of the value in the following ways:

A simple value is wrapped in "Headers '[hdr]":

>>> let example0 = addHeader 5 "hi" :: Headers '[Header "someheader" Int] String;
>>> getHeaders example0
[("someheader","5")]

A value that already has a header has its new header *prepended* to the existing list:

>>> let example1 = addHeader 5 "hi" :: Headers '[Header "someheader" Int] String;
>>> let example2 = addHeader True example1 :: Headers '[Header "1st" Bool, Header "someheader" Int] String
>>> getHeaders example2
[("1st","true"),("someheader","5")]

Note that while in your handlers type annotations are not required, since the type can be inferred from the API type, in other cases you may find yourself needing to add annotations.

lookupResponseHeader :: HasResponseHeader h a headers => Headers headers r -> ResponseHeader h a

Look up a specific ResponseHeader, without having to know what position it is in the HList.

>>> let example1 = addHeader 5 "hi" :: Headers '[Header "someheader" Int] String
>>> let example2 = addHeader True example1 :: Headers '[Header "1st" Bool, Header "someheader" Int] String
>>> lookupResponseHeader example2 :: ResponseHeader "someheader" Int
Header 5

>>> lookupResponseHeader example2 :: ResponseHeader "1st" Bool
Header True

Usage of this function relies on an explicit type annotation of the header to be looked up. This can be done with type annotations on the result, or with an explicit type application. In this example, the type of header value is determined by the type-inference, we only specify the name of the header:

>>> :set -XTypeApplications
>>> case lookupResponseHeader @"1st" example2 of { Header b -> b ; _ -> False }
True

Since: 0.15

noHeader :: AddHeader h v orig new => orig -> new

Deliberately do not add a header to a value.

>>> let example1 = noHeader "hi" :: Headers '[Header "someheader" Int] String
>>> getHeaders example1
[]

Untyped endpoints

data Raw

Endpoint for plugging in your own Wai Applications.

The given Application will get the request as received by the server, potentially with a modified (stripped) pathInfo if the Application is being routed with :>.

In addition to just letting you plug in your existing WAI Applications, this can also be used with functions from Servant.Server.StaticFiles to serve static files stored in a particular directory on your filesystem

data RawM

Variant of Raw that lets you access the underlying monadic context to process the request.

Plugging in a wai Application, serving directories

FromHttpApiData and ToHttpApiData

class FromHttpApiData a where

Parse value from HTTP API data.

WARNING: Do not derive this using DeriveAnyClass as the generated instance will loop indefinitely.

Minimal complete definition

parseUrlPiece | parseQueryParam

Methods

parseUrlPiece :: Text -> Either Text a

Parse URL path piece.

parseHeader :: ByteString -> Either Text a

Parse HTTP header value.

parseQueryParam :: Text -> Either Text a

Parse query param value.

class ToHttpApiData a where

Convert value to HTTP API data.

WARNING: Do not derive this using DeriveAnyClass as the generated instance will loop indefinitely.

Minimal complete definition

toUrlPiece | toQueryParam

Methods

toUrlPiece :: a -> Text

Convert to URL path piece.

toEncodedUrlPiece :: a -> Builder

Convert to a URL path piece, making sure to encode any special chars. The default definition uses urlEncodeBuilder False but this may be overriden with a more efficient version.

toHeader :: a -> ByteString

Convert to HTTP header value.

toQueryParam :: a -> Text

Convert to query param value.

toEncodedQueryParam :: a -> Builder

Convert to URL query param, The default definition uses urlEncodeBuilder True but this may be overriden with a more efficient version.

Since: http-api-data-0.5.1

Classes and instances for types that can be converted to and from HTTP API data.

Experimental modules

data AuthProtect (tag :: k)

A generalized Authentication combinator. Use this if you have a non-standard authentication technique.

NOTE: THIS API IS EXPERIMENTAL AND SUBJECT TO CHANGE.

General Authentication

Links

data URI

Represents a general universal resource identifier using its component parts.

For example, for the URI

  foo://anonymous@www.haskell.org:42/ghc?query#frag

the components are:

Constructors

URI

Fields

uriScheme :: String

foo:

uriAuthority :: Maybe URIAuth

//anonymous@www.haskell.org:42

uriPath :: String

/ghc

uriQuery :: String

?query

uriFragment :: String

#frag

class HasLink endpoint where

Construct a toLink for an endpoint.

Associated Types

type MkLink endpoint (a :: *)

Methods

toLink

Arguments

:: (Link -> a)

-> Proxy endpoint
The API endpoint you would like to point to

-> Link

-> MkLink endpoint a

type family IsElem endpoint api :: Constraint where ...

Closed type family, check if endpoint is within api. Uses IsElem' if it exhausts all other options.

>>> ok (Proxy :: Proxy (IsElem ("hello" :> Get '[JSON] Int) SampleAPI))
OK

>>> ok (Proxy :: Proxy (IsElem ("bye" :> Get '[JSON] Int) SampleAPI))
...
... Could not ...
...

An endpoint is considered within an api even if it is missing combinators that don't affect the URL:

>>> ok (Proxy :: Proxy (IsElem (Get '[JSON] Int) (Header "h" Bool :> Get '[JSON] Int)))
OK

>>> ok (Proxy :: Proxy (IsElem (Get '[JSON] Int) (ReqBody '[JSON] Bool :> Get '[JSON] Int)))
OK

N.B.:* IsElem a b can be seen as capturing the notion of whether the URL represented by a would match the URL represented by b, *not* whether a request represented by a matches the endpoints serving b (for the latter, use IsIn).

Equations

IsElem e (sa :<|> sb) = Or (IsElem e sa) (IsElem e sb)

IsElem (e :> sa) (e :> sb) = IsElem sa sb

IsElem sa (Header sym x :> sb) = IsElem sa sb

IsElem sa (ReqBody y x :> sb) = IsElem sa sb

IsElem (CaptureAll z y :> sa) (CaptureAll x y :> sb) = IsElem sa sb

IsElem (Capture z y :> sa) (Capture x y :> sb) = IsElem sa sb

IsElem sa (QueryParam x y :> sb) = IsElem sa sb

IsElem sa (QueryParams x y :> sb) = IsElem sa sb

IsElem sa (QueryFlag x :> sb) = IsElem sa sb

IsElem sa (Fragment x :> sb) = IsElem sa sb

IsElem (Verb m s ct typ) (Verb m s ct' typ) = IsSubList ct ct'

IsElem e e = ()

IsElem e a = IsElem' e a

type family IsElem' a s :: Constraint

You may use this type family to tell the type checker that your custom type may be skipped as part of a link. This is useful for things like QueryParam that are optional in a URI and do not affect them if they are omitted.

>>> data CustomThing
>>> type instance IsElem' e (CustomThing :> s) = IsElem e s

Note that IsElem is called, which will mutually recurse back to IsElem' if it exhausts all other options again.

Once you have written a HasLink instance for CustomThing you are ready to go.

data Link

A safe link datatype. The only way of constructing a Link is using safeLink, which means any Link is guaranteed to be part of the mentioned API.

safeLink

Arguments

:: forall endpoint api. (IsElem endpoint api, HasLink endpoint)

=> Proxy api
The whole API that this endpoint is a part of

-> Proxy endpoint
The API endpoint you would like to point to

-> MkLink endpoint Link

Create a valid (by construction) relative URI with query params.

This function will only typecheck if endpoint is part of the API api

Type-safe internal URIs

Re-exports

type family If (cond :: Bool) (tru :: k) (fls :: k) :: k where ...

Type-level If. If True a b ==> a; If False a b ==> b

Equations

If 'True (tru :: k) (fls :: k) = tru

If 'False (tru :: k) (fls :: k) = fls

data SBool (b :: Bool) where

Constructors

STrue :: SBool 'True

SFalse :: SBool 'False

class SBoolI (b :: Bool) where

Methods

sbool :: SBool b