Servant.API.ResponseHeaders

Description

This module provides facilities for adding headers to a response.

>>> let headerVal = addHeader "some-url" 5 :: Headers '[Header "Location" String] Int

The value is added to the header specified by the type (Location in the example above).

Documentation

data Headers ls a

Response Header objects. You should never need to construct one directly. Instead, use addHeader.

Constructors

Headers

Fields

getResponse :: a

The underlying value of a Headers

getHeadersHList :: HList ls

HList of headers.

class AddHeader h v orig new | h v orig -> new, new -> h, new -> v, new -> orig where

Methods

addHeader

Arguments

:: v

-> orig

-> new
N.B.: The same header can't be added multiple times

class BuildHeadersTo hs where

Methods

buildHeadersTo :: [Header] -> HList hs

Note: if there are multiple occurences of a header in the argument, the values are interspersed with commas before deserialization (see RFC2616 Sec 4.2)

class GetHeaders ls where

Methods

getHeaders :: ls -> [Header]

type family HeaderValMap f xs

Equations

HeaderValMap f `[]` = `[]`

HeaderValMap f (Header h x : xs) = Header h (f x) : HeaderValMap f xs

data HList a where

Constructors

HNil :: HList `[]`

HCons :: Header h x -> HList xs -> HList (Header h x : xs)