Servant.API.ResponseHeaders

Description

This module provides facilities for adding headers to a response.

>>> let headerVal = addHeader "some-url" 5 :: Headers '[Header "Location" String] Int

The value is added to the header specified by the type (Location in the example above).

Documentation

data Headers ls a

Response Header objects. You should never need to construct one directly. Instead, use addOptionalHeader.

Constructors

Headers

Fields

getResponse :: a

The underlying value of a Headers

getHeadersHList :: HList ls

HList of headers.

data ResponseHeader (sym :: Symbol) a

Constructors

Header a

MissingHeader

UndecodableHeader ByteString

class AddHeader h v orig new | h v orig -> new, new -> h, new -> v, new -> orig

Minimal complete definition

addOptionalHeader

addHeader :: AddHeader h v orig new => v -> orig -> new

addHeader adds a header to a response. Note that it changes the type of the value in the following ways:

A simple value is wrapped in "Headers '[hdr]":

>>> let example1 = addHeader 5 "hi" :: Headers '[Header "someheader" Int] String;
>>> getHeaders example1
[("someheader","5")]

A value that already has a header has its new header *prepended* to the existing list:

>>> let example1 = addHeader 5 "hi" :: Headers '[Header "someheader" Int] String;
>>> let example2 = addHeader True example1 :: Headers '[Header "1st" Bool, Header "someheader" Int] String
>>> getHeaders example2
[("1st","true"),("someheader","5")]

Note that while in your handlers type annotations are not required, since the type can be inferred from the API type, in other cases you may find yourself needing to add annotations.

noHeader :: AddHeader h v orig new => orig -> new

Deliberately do not add a header to a value.

>>> let example1 = noHeader "hi" :: Headers '[Header "someheader" Int] String
>>> getHeaders example1
[]

class HasResponseHeader h a headers

Minimal complete definition

hlistLookupHeader

lookupResponseHeader :: HasResponseHeader h a headers => Headers headers r -> ResponseHeader h a

Look up a specific ResponseHeader, without having to know what position it is in the HList.

>>> let example1 = addHeader 5 "hi" :: Headers '[Header "someheader" Int] String
>>> let example2 = addHeader True example1 :: Headers '[Header "1st" Bool, Header "someheader" Int] String
>>> lookupResponseHeader example2 :: ResponseHeader "someheader" Int
Header 5

>>> lookupResponseHeader example2 :: ResponseHeader "1st" Bool
Header True

Usage of this function relies on an explicit type annotation of the header to be looked up. This can be done with type annotations on the result, or with an explicit type application. In this example, the type of header value is determined by the type-inference, we only specify the name of the header:

>>> :set -XTypeApplications
>>> case lookupResponseHeader @"1st" example2 of { Header b -> b ; _ -> False }
True

Since: 0.15

class BuildHeadersTo hs where

Methods

buildHeadersTo :: [Header] -> HList hs

Note: if there are multiple occurrences of a header in the argument, the values are interspersed with commas before deserialization (see RFC2616 Sec 4.2)

class GetHeaders ls where

Methods

getHeaders :: ls -> [Header]

class GetHeaders' hs

Auxiliary class for GetHeaders (Headers hs a) instance

Minimal complete definition

getHeaders'

type family HeaderValMap (f :: * -> *) (xs :: [*]) where ...

Equations

HeaderValMap f '[] = '[]

HeaderValMap f (Header h x ': xs) = Header h (f x) ': HeaderValMap f xs

data HList a where

Constructors

HNil :: HList '[]

HCons :: ResponseHeader h x -> HList xs -> HList (Header h x ': xs)